+++ service/http/config.go
@@ -9,9 +9,6 @@ import (
 
 // Config configures RoadRunner HTTP server.
 type Config struct {
-	// Enable enables http service.
-	Enable bool
-
 	// Address and port to handle as http server.
 	Address string
 
@@ -27,20 +24,26 @@ type Config struct {
 
 // Hydrate must populate Config values using given Config source. Must return error if Config is not valid.
 func (c *Config) Hydrate(cfg service.Config) error {
+	if c.Workers == nil {
+		c.Workers = &roadrunner.ServerConfig{}
+	}
+
+	if c.Uploads == nil {
+		c.Uploads = &UploadsConfig{}
+	}
+
+	c.Uploads.InitDefaults()
+	c.Workers.InitDefaults()
+
 	if err := cfg.Unmarshal(c); err != nil {
 		return err
 	}
 
-	if !c.Enable {
-		return nil
-	}
+	c.Workers.UpscaleDurations()
 
 	if err := c.Valid(); err != nil {
 		return err
 	}
-
-	c.Workers.SetDefaults()
-	c.Workers.UpscaleDurations()
 
 	return nil
 }