+++ service/http/config_test.go
@@ -31,7 +31,6 @@ func Test_Config_Hydrate_Error2(t *testing.T) {
 
 func Test_Config_Valid(t *testing.T) {
 	cfg := &Config{
-		Enable:     true,
 		Address:    ":8080",
 		MaxRequest: 1024,
 		Uploads: &UploadsConfig{
@@ -39,7 +38,7 @@ func Test_Config_Valid(t *testing.T) {
 			Forbid: []string{".go"},
 		},
 		Workers: &roadrunner.ServerConfig{
-			Command: "php php-src/tests/client.php echo pipes",
+			Command: "php tests/client.php echo pipes",
 			Relay:   "pipes",
 			Pool: &roadrunner.Config{
 				NumWorkers:      1,
@@ -54,11 +53,10 @@ func Test_Config_Valid(t *testing.T) {
 
 func Test_Config_NoUploads(t *testing.T) {
 	cfg := &Config{
-		Enable:     true,
 		Address:    ":8080",
 		MaxRequest: 1024,
 		Workers: &roadrunner.ServerConfig{
-			Command: "php php-src/tests/client.php echo pipes",
+			Command: "php tests/client.php echo pipes",
 			Relay:   "pipes",
 			Pool: &roadrunner.Config{
 				NumWorkers:      1,
@@ -73,7 +71,6 @@ func Test_Config_NoUploads(t *testing.T) {
 
 func Test_Config_NoWorkers(t *testing.T) {
 	cfg := &Config{
-		Enable:     true,
 		Address:    ":8080",
 		MaxRequest: 1024,
 		Uploads: &UploadsConfig{
@@ -87,7 +84,6 @@ func Test_Config_NoWorkers(t *testing.T) {
 
 func Test_Config_NoPool(t *testing.T) {
 	cfg := &Config{
-		Enable:     true,
 		Address:    ":8080",
 		MaxRequest: 1024,
 		Uploads: &UploadsConfig{
@@ -95,7 +91,7 @@ func Test_Config_NoPool(t *testing.T) {
 			Forbid: []string{".go"},
 		},
 		Workers: &roadrunner.ServerConfig{
-			Command: "php php-src/tests/client.php echo pipes",
+			Command: "php tests/client.php echo pipes",
 			Relay:   "pipes",
 			Pool: &roadrunner.Config{
 				NumWorkers:      0,
@@ -110,7 +106,6 @@ func Test_Config_NoPool(t *testing.T) {
 
 func Test_Config_DeadPool(t *testing.T) {
 	cfg := &Config{
-		Enable:     true,
 		Address:    ":8080",
 		MaxRequest: 1024,
 		Uploads: &UploadsConfig{
@@ -118,7 +113,7 @@ func Test_Config_DeadPool(t *testing.T) {
 			Forbid: []string{".go"},
 		},
 		Workers: &roadrunner.ServerConfig{
-			Command: "php php-src/tests/client.php echo pipes",
+			Command: "php tests/client.php echo pipes",
 			Relay:   "pipes",
 		},
 	}
@@ -128,7 +123,6 @@ func Test_Config_DeadPool(t *testing.T) {
 
 func Test_Config_InvalidAddress(t *testing.T) {
 	cfg := &Config{
-		Enable:     true,
 		Address:    "",
 		MaxRequest: 1024,
 		Uploads: &UploadsConfig{
@@ -136,7 +130,7 @@ func Test_Config_InvalidAddress(t *testing.T) {
 			Forbid: []string{".go"},
 		},
 		Workers: &roadrunner.ServerConfig{
-			Command: "php php-src/tests/client.php echo pipes",
+			Command: "php tests/client.php echo pipes",
 			Relay:   "pipes",
 			Pool: &roadrunner.Config{
 				NumWorkers:      1,