+++ socket_factory_test.go
@@ -19,7 +19,7 @@ func Test_Tcp_Start(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "tcp")
+	cmd := exec.Command("php", "tests/client.php", "echo", "tcp")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	assert.NoError(t, err)
@@ -41,7 +41,7 @@ func Test_Tcp_StartCloseFactory(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "tcp")
+	cmd := exec.Command("php", "tests/client.php", "echo", "tcp")
 
 	f := NewSocketFactory(ls, time.Minute)
 	defer f.Close()
@@ -67,7 +67,7 @@ func Test_Tcp_StartError(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 	cmd.Start()
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
@@ -85,7 +85,7 @@ func Test_Tcp_Failboot(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/failboot.php")
+	cmd := exec.Command("php", "tests/failboot.php")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	assert.Nil(t, w)
@@ -103,7 +103,7 @@ func Test_Tcp_Timeout(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/slow-client.php", "echo", "tcp", "200", "0")
+	cmd := exec.Command("php", "tests/slow-client.php", "echo", "tcp", "200", "0")
 
 	w, err := NewSocketFactory(ls, time.Millisecond*100).SpawnWorker(cmd)
 	assert.Nil(t, w)
@@ -121,7 +121,7 @@ func Test_Tcp_Invalid(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/invalid.php")
+	cmd := exec.Command("php", "tests/invalid.php")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	assert.Error(t, err)
@@ -138,7 +138,7 @@ func Test_Tcp_Broken(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "broken", "tcp")
+	cmd := exec.Command("php", "tests/client.php", "broken", "tcp")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	go func() {
@@ -165,7 +165,7 @@ func Test_Tcp_Echo(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "tcp")
+	cmd := exec.Command("php", "tests/client.php", "echo", "tcp")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	go func() {
@@ -195,7 +195,7 @@ func Test_Unix_Start(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "unix")
+	cmd := exec.Command("php", "tests/client.php", "echo", "unix")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	assert.NoError(t, err)
@@ -220,7 +220,7 @@ func Test_Unix_Failboot(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/failboot.php")
+	cmd := exec.Command("php", "tests/failboot.php")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	assert.Nil(t, w)
@@ -240,7 +240,7 @@ func Test_Unix_Timeout(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/slow-client.php", "echo", "unix", "200", "0")
+	cmd := exec.Command("php", "tests/slow-client.php", "echo", "unix", "200", "0")
 
 	w, err := NewSocketFactory(ls, time.Millisecond*100).SpawnWorker(cmd)
 	assert.Nil(t, w)
@@ -260,7 +260,7 @@ func Test_Unix_Invalid(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/invalid.php")
+	cmd := exec.Command("php", "tests/invalid.php")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	assert.Error(t, err)
@@ -279,7 +279,7 @@ func Test_Unix_Broken(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "broken", "unix")
+	cmd := exec.Command("php", "tests/client.php", "broken", "unix")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	go func() {
@@ -308,7 +308,7 @@ func Test_Unix_Echo(t *testing.T) {
 		t.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "unix")
+	cmd := exec.Command("php", "tests/client.php", "echo", "unix")
 
 	w, err := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	go func() {
@@ -336,7 +336,7 @@ func Benchmark_Tcp_SpawnWorker_Stop(b *testing.B) {
 
 	f := NewSocketFactory(ls, time.Minute)
 	for n := 0; n < b.N; n++ {
-		cmd := exec.Command("php", "php-src/tests/client.php", "echo", "tcp")
+		cmd := exec.Command("php", "tests/client.php", "echo", "tcp")
 
 		w, _ := f.SpawnWorker(cmd)
 		go func() {
@@ -357,7 +357,7 @@ func Benchmark_Tcp_Worker_ExecEcho(b *testing.B) {
 		b.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "tcp")
+	cmd := exec.Command("php", "tests/client.php", "echo", "tcp")
 
 	w, _ := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	go func() {
@@ -386,7 +386,7 @@ func Benchmark_Unix_SpawnWorker_Stop(b *testing.B) {
 
 	f := NewSocketFactory(ls, time.Minute)
 	for n := 0; n < b.N; n++ {
-		cmd := exec.Command("php", "php-src/tests/client.php", "echo", "unix")
+		cmd := exec.Command("php", "tests/client.php", "echo", "unix")
 
 		w, _ := f.SpawnWorker(cmd)
 		go func() {
@@ -411,7 +411,7 @@ func Benchmark_Unix_Worker_ExecEcho(b *testing.B) {
 		b.Skip("socket is busy")
 	}
 
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "unix")
+	cmd := exec.Command("php", "tests/client.php", "echo", "unix")
 
 	w, _ := NewSocketFactory(ls, time.Minute).SpawnWorker(cmd)
 	go func() {