+++ src/Exception/RoadRunnerException.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * High-performance PHP process supervisor and load balancer written in Go
+ *
+ * @author    Wolfy-J
+ */
+
+namespace Spiral\RoadRunner\Exception;
+
+class RoadRunnerException extends \Spiral\RoadRunner\Exceptions\RoadRunnerException
+{
+
+}