+++ worker_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func Test_GetState(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 
 	w, err := NewPipeFactory().SpawnWorker(cmd)
 	go func() {
@@ -23,7 +23,7 @@ func Test_GetState(t *testing.T) {
 }
 
 func Test_Kill(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 
 	w, err := NewPipeFactory().SpawnWorker(cmd)
 	go func() {
@@ -39,7 +39,7 @@ func Test_Kill(t *testing.T) {
 }
 
 func Test_Echo(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 
 	w, _ := NewPipeFactory().SpawnWorker(cmd)
 	go func() {
@@ -58,7 +58,7 @@ func Test_Echo(t *testing.T) {
 }
 
 func Test_BadPayload(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 
 	w, _ := NewPipeFactory().SpawnWorker(cmd)
 	go func() {
@@ -75,16 +75,16 @@ func Test_BadPayload(t *testing.T) {
 }
 
 func Test_NotStarted_String(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 
 	w, _ := newWorker(cmd)
-	assert.Contains(t, w.String(), "php php-src/tests/client.php echo pipes")
+	assert.Contains(t, w.String(), "php tests/client.php echo pipes")
 	assert.Contains(t, w.String(), "inactive")
 	assert.Contains(t, w.String(), "numExecs: 0")
 }
 
 func Test_NotStarted_Exec(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 
 	w, _ := newWorker(cmd)
 
@@ -97,7 +97,7 @@ func Test_NotStarted_Exec(t *testing.T) {
 }
 
 func Test_String(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 
 	w, _ := NewPipeFactory().SpawnWorker(cmd)
 	go func() {
@@ -105,13 +105,13 @@ func Test_String(t *testing.T) {
 	}()
 	defer w.Stop()
 
-	assert.Contains(t, w.String(), "php php-src/tests/client.php echo pipes")
+	assert.Contains(t, w.String(), "php tests/client.php echo pipes")
 	assert.Contains(t, w.String(), "ready")
 	assert.Contains(t, w.String(), "numExecs: 0")
 }
 
 func Test_Echo_Slow(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/slow-client.php", "echo", "pipes", "10", "10")
+	cmd := exec.Command("php", "tests/slow-client.php", "echo", "pipes", "10", "10")
 
 	w, _ := NewPipeFactory().SpawnWorker(cmd)
 	go func() {
@@ -130,7 +130,7 @@ func Test_Echo_Slow(t *testing.T) {
 }
 
 func Test_Broken(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "broken", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "broken", "pipes")
 
 	w, err := NewPipeFactory().SpawnWorker(cmd)
 	go func() {
@@ -146,7 +146,7 @@ func Test_Broken(t *testing.T) {
 }
 
 func Test_OnStarted(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "broken", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "broken", "pipes")
 	assert.Nil(t, cmd.Start())
 
 	w, err := newWorker(cmd)
@@ -157,7 +157,7 @@ func Test_OnStarted(t *testing.T) {
 }
 
 func Test_Error(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "error", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "error", "pipes")
 
 	w, _ := NewPipeFactory().SpawnWorker(cmd)
 	go func() {
@@ -174,7 +174,7 @@ func Test_Error(t *testing.T) {
 }
 
 func Test_NumExecs(t *testing.T) {
-	cmd := exec.Command("php", "php-src/tests/client.php", "echo", "pipes")
+	cmd := exec.Command("php", "tests/client.php", "echo", "pipes")
 
 	w, _ := NewPipeFactory().SpawnWorker(cmd)
 	go func() {